+++ Wms/Application/Inventory/Movements/InventoryMovementListItem.cs
@@ -14,7 +14,7 @@ public class InventoryMovementListItem
     public StorageLocation? SourceStorageLocation => Movement.SourceStorageLocation;
     public StorageLocation? DestinationStorageLocation => Movement.DestinationStorageLocation;
     public StockKeepingUnit? StockKeepingUnit => Movement.StockKeepingUnit;
-    public double Quantity => Movement.Quantity;
+    public decimal Quantity => Movement.Quantity;
     public double? WeightKg => Movement.WeightKg;
     public RecorderType RecorderType => Movement.RecorderType;
     public Guid? RecorderId => Movement.RecorderId;
